[1/2] Made EmuTime/Duration/Clock constexpr
This allows to performs operations on these objects (also) during compile-time
(rather than only during run-time). Including construction of these objects
during compile time.

We now use this for example in V9990CmdEngine to replace a table of cycle
counts with a table of EmuDuration objects. Because the only thing that was
done with those cycle counts was converting them to an EmuDuration (at
run-time), now that conversion is already done during compile-time.