Handle symlink creation failing on Windows 10
Apparently Python 3 can create symlinks under Windows 10, unlike
Python 2.7, but it needs high privileges to do so. I don't think
our build should require special privileges, so I'll be working
around it instead.

The symlinks are in the macOS build support for SDL2_ttf, so there
is no harm in not creating the symlinks on Windows.