Fix crash in Keyboard with ColecoVision machine
The modifierPosForMatrix[] array in the Keyboard code is not filled in for the
ColecoVision machine (meaning the elements get default initialized to the
'invalid' KeyMatrixPosition value). I have no idea if this is intentional or
not. But because of this (before this patch) the code would crash in
updateKeyMatrix() when e.g. ALT+SPACE was pressed when emulating the
ColecoVision machine.