Experimental, not-yet-finished: add support for Yamaha FD-03 FDC
This patch is for a large part developed by, or based on information provided
by, Arjen Zeilemaker. Thanks a lot!

Arjen claimed this patch (partly) started to work, but it doesn't work for me.
Possibly I'm doing something wrong, or possibly I screwed up something while
integrating his patches. I'm commiting it anyway to make it easier to
coordinate working on this.

The main changes are:
- Extended DriveMultiplexer API with the functions:
    isDiskInserted(), diskChanged(), peekDiskChange()
  These functions already existed on the DiskDrive class but not yet on
  DriveMultiplexer.
- Added a 'trackMode' to RealDrive:
  Normally one step of the disk drive read/write head corresponds to a movement
  of 1 track on the disk. Though for the Yamaha FD-03 the head needs to take 4
  steps to move over a distance of only 1 disk track.
- WD2793 emulation:
  Support multi-sector read/write commands.
  Read-address command should fill the sector register with the track number
  found in the sector header.
- New class YamahaFDC that implements the YamahaFDC FD-03.