Replace snappy with lz4 compression
The reverse subsystem takes regular in-memory snapshots of the whole MSX state.
Part of that state is the content of the different RAM blocks. Most of those
blocks are stored as a delta against a reference block (from an earlier
snapshot). But the reference blocks are full copies of the original MSX RAM.

To save space (all but the last) reference blocks are compressed. Before this
patch we used the 'snappy' compression algorithm for this:
  https://github.com/google/snappy
now we use the lz4 algorithm:
  https://github.com/lz4/lz4
Because compression is done while MSX emulation is running(*), it's important
that it runs very fast. Achieving the best compression ratio is less important.
Both algorithms satisfy this criteria.

In the past I had already compared lz4 with snappy and at the time snappy
seemed better. Though more and more projects use lz4 rather than snappy. And
many benchmarks you find on the internet also favor lz4. So therefor I decided
to reevaluate. Today my tests show that:
 - lz4 compression runs ~20% faster compared to snappy.
 - The lz4 produced output is 4% smaller than the snappy output.

Possible explanations for why I'm getting the opposite test results now are:
 - I'm using a newer version of the lz4 code.
 - I'm not testing on the same host machine.
 - Compression runs in a different context now versus back then (before/after
   the delta-block refactoring). Because of this the to-be-compressed data
   *might* be present in a different CPU cache level??

I did not explicitly measure decompression speeds, because those are less
important in our context. (But various benchmarks on the internet show that lz4
decompression is faster than snappy). Note that (for both lz4 and snappy) we
use modified decompression routines. The original routines had various safety
check to deal with malicious input data. But in openMSX the only input we give
to the compression routine is data that was previously produced by the
compression routine. That data always stays in-memory (it's not part of some
file format).

(*) Idea for the future: run the compression on a different thread. This only
    became possible after the delta-block refactoring.