Made VDP.commandExecuting probe more accurate
This is a follow-up on 29eeaffbffe30c2.

Before the existence of the VDP.commandExecuting probe it was not needed to
predict in advance when a VDP command would finish. I overlooked this detail
yesterday when implementing that probe. A consequence is that the probe will
only trigger when the emulation synchronizes with the VDPCmdEngine (e.g. when
polling the CE bit or when rendering a frame, ...). Though at that point CPU
emulation (the main emulation time driver) may have advanced well past the time
of the end of the command, and then it will appear as if the probe triggered
too late.

To make the probe more accurate we will now more actively sync with (estimates
for) the end of a command. But only when the probe is actually in use.