[2/3] split cmdKeyMatrix into cmdKeyMatrix + typeKeyMatrix
Before this patch we (only) had separate keyboard matrices for what the user is
typing via the host keyboard (live or via replay) and for what is being typed
via commands. Though there are two sets of commands that influence the keyboard
matrix namely 'type' and 'keymatrixup/down', and those still interfere.

So this patch splits the matrix for the commands into two.