[3/3] exclusive type or user matrix
Before this patch, the keyboard matrices for the type command and for the
'normal' user input were simply mixed together. This means that user input
while a type command is running would garble up that type command. Also the
default keybinding 'CTRl+META+V' for 'paste' has the CTRL modifier pressed
and that often messed up the first few characters of the pasted text.

This patch changes the matrix combination code to exclusively look at either
the typeKeyMatrix (when a type command is running) or the userKeyMatrix. In
both cases we still unconditionally mix that with the cmdKeyMatrix (controlled
via keymatrixup/down commands).