Two tiny tweaks to setting breakpoints/conditions
When the callee will unconditionally make a copy of a function parameter (IOW a
sink-parameter), it can be cheaper to pass that parameter by-value (and
std::move() it) instead of passing it by const-reference.

In practice this won't make much difference because currently BreakPoint and
DebugCondition objects are as cheap to copy as to move. Though the new code is
more idiomatic, and more robust against internal changes in BreakPoint /
DebugCondition.

Another tiny tweak in MSXCPUInterface::doStep(): an empty string for the
condition evaluates a bit faster than the condition "true" (the former bypasses
the Tcl interpreter). Not that speed really matters on that code path.