Add PortAudio to 3rdparty build system
I excluded pa_hostapi_skeleton from the build because the build breaks
when the object for this file is written without an output directory
for it being created in an out-of-tree build. Since this is a skeleton
for new host drivers, the only point in compiling it is for PortAudio
developers to verify that it still compiles, but when using the library
there is no need for it.

While some parts of PortAudio's configure script use pkg-config,
the part that probes ALSA instead using a hardcoded flag, which is
insufficient when linking statically (missing -lm -lpthread).
I hardcoded the use of pkg-config since we rely on pkg-config
already for the 3rdparty build.