Bug#1183 [2/4]: factor out small helper functions
Try to eliminate duplicated code as much as possible. (And rely on the compiler
inlining these helpers to get the same performance back).

For example the code that draws a (tile) pixel was duplicated ~10x. Currently
that code is a one-liner, so that duplication doesn't matter. In the next
patch, we will change this drawing (and specialize it for foregound and
background). And I prefer to not have to duplicated those changes 10x.

So this patch factors out some code, but doesn't make any functional changes
yet (apart from some minor tweaks).