Bug#1183 [3/4]: change render algorithm
This patch does the actual bug fix. See
  https://github.com/openMSX/openMSX/issues/1183
for a very detailed bug report and test case. Thanks GhostwriterP!

The old render algorithm worked like this:
- Fill buffer with backdrop color.
- Overdraw with background tiles (skip transparent pixels).
- Overdraw with back sprites, from low to high priority (thus higher priority
  sprites overdraw lower priority sprites).
- Overdraw with foregound tiles.
- Overdraw with front sprites (again from low to high priority).

On the other hand the real hardware renders pixels from left to right. Thus per
pixel it determines which of the 5 planes is visible, and which is the highest
priority sprite within the sprite-plane(s). Our very first P1 renderer also
worked like this, but it was _very_ slow compared to the above algorithm.
Possibly it could be tweaked, but especially sprites rendering requires a lot
more (256x more) check when done per pixel, than when done per line.

So the new render algorithm remains a line-based renderer. But, to the best of
my knowledge, it should produce the same result as the real hardware's
pixel-based renderer. It goes like this:
- Draw combined backdrop + backgound tiles. (This combination could also be
  done in the old algorithm, it's an optimization I only thought of now).
- Overdraw with the forground tiles. For each pixel remember wheter it was
  transparent or not (thus whether the pixel contains a background tile (or
  backdrop color) or foreground tile).
  (Note that the back sprite plane (which goes _between_ the two tile planes)
  has not yey been drawn).
- In the last step we draw both sprite planes at once. Now we draw sprites from
  high to low priority. For each non-transparent sprites pixel we:
  * Check if another sprites was already drawn on this pixel (sprites don't
    overdraw other sprites anymore). If so go to next pixel.
  * Mark this pixel as containing a sprite (even though next step can still
    skip!).
  * Draw the pixel for front-sprite or for back-sprite when the pixel does not
    contain a forground tile.