Bug#1183 [4/4]: speedup backdrop+background drawing
The first pass of the new render algorithm simultaneously draws the backgound
tiles and backdrop color. Per pixel it basically does an operation like:
    // index in range 0..15
    *buffer = index ? palette[index] : backDropColor;

So entry-0 of the palette[] array is never actually used. We can exploit this
by _temporarily_ replacing that entry with 'backDropColor'. Then the operation
simplifies to:
    *buffer = palette[index];