[2/2] factor out common code between P1 and P2 renderers
See the bug report for details:
   https://github.com/openMSX/openMSX/issues/1183
D15CODE wondered why P1 and P2 don't share (more) code. I totally agree, having
this stuff shared would also have prevented me from fixing a bug in P1 but then
forgetting about P2.

Though while working on this patch I remember again why this code wasn't shared
from the start: from a distance it indeed looks like the only difference is the
screen width (256 vs 512) (and of course that P1 has two instead of one tile
plane). Though when you look in more detail, you'll see many more differences:
 - For tiles: address calculations for name-table, and pattern-table.
 - For sprites: address calculation for pattern-table.
 - Logical to physical VRAM mapping is different in P1 vs P2.
Combined all these small difference make that actually none of the sub functions
were the same for P1 and P2. And the differences are not cheaply abstracted via
a function parameter. (And performance is important for this code).

Nevertheless the structure of the algorithm is the same for P1 and P2, so it's
indeed desirable to only have to express this structure once. Today I thought
of using a Policy based design to:
 - Express the general structure only once.
 - Extract the differences into (many) small sub functions.
 - Rely on c++ template instantiation to generate code that qua performance is
   on par with the original non-shared code.