[2/2] Allow short EmuDurations to be stored more compactly
Internally an EmuDuration object stores a (single) 64-bit value. When this
object represents a short duration (a few milliseconds or even microseconds)
then the numerical value stored in that 64-bit value could also fit in a 32, 16
or even an 8 bit value.

This patch introduces 3 EmuDuration storage classes: EmuDuration{32,16,8}, for
when the value would fit in respectively 32, 16 or 8 bits. The only operations
that are available on these classes are conversion from and to an actual
EmuDuration object. So it's really only meant to be a more compact storage
vehicle.

In the previous patch we changed tables of (32 bit) integers into tables of
EmuDuration objects. This doubled the size of those tables. The values in those
tables represent the time needed to execute one operation in a V9990 command,
so these are all fairly short. It turns out they all fit in 16-bit. So by
storing them as EmuDuration16 objects we reduce the size of the tables by a
factor 4, thus also half the size of the original integer tables.