[1/2] Use c++17 inline variables
Pre-c++17 static member variable had to be:
- Declared in the class definition (typically in a .hh file).
- And defined in exactly one .cc file.

For example:
    -- myfile.hh --
    ...
    class Foo {
        ...
        static int var;
        ...
    };

    -- myfile.cc --
    ...
    int Foo::var = 42;
    ...

C++17 introduces 'inline variables' to simplify this pattern. The example then
becomes:
    -- myfile.hh --
    ...
    class Foo {
        ...
        static inline int var = 42;   // no need to repeat in .cc file
        ...
    };

For inline functions the keyword 'inline' has long lost it's meaning of
"include one function into another (saving the call/return overhead, and
enabling further optimization opportunities)". The remaining meaning of
'inline' is: "allow this definition to appear in multiple translation units,
keep only one definition after linking" (typically because an inline function
is defined in a .hh file and that file is #included in several translation
units). It's this latter meaning that applies to 'inline' variables. For more
details see:
    https://en.cppreference.com/w/cpp/language/inline

Note: static member variables that are also constant are handled in the next
patch.