Fixed bugs in gl_{vec,mat} unittest
As an experiment I enabled the address-sanitizer for the unittests. This did
uncover some bugs in the test-code (the main code itself was fine).

The bugs were all copy-paste errors:
    vec4 v4(4, 5, 6, 7);
    vec4 w4(v4);
    CHECK(v4 == w4);
    v3[3] = 0; w4 = v4; // <-- should have been 'v4[3] = 0'
    CHECK(v4 == w4);
I've rewritten the tests is such a way that this kind of mistake is impossible:
- Split the test in smaller chunks, so that the wrong variable is no longer in
  scope.
- Add a CHECK that the variables are no longer equal after the assignment.

Because this experiment found some actual bugs, I kept the sanitizer enabled in
this patch. But we have to evaluate what effect this has on the automatic build
loop (the compilation step is quite a bit heavier).