util: added c++17 [[nodiscard]] annotations
Added [[nodiscard]] annotations on the functions/methods in the util/
subdirectory.

This is a new feature in c++17. It instructs the compiler to generate a warning
when a caller ignores the return value of the call.

Motivation: at work we started using these annotations and they uncovered some
bugs. For example:
 - using operator+ when operator+= was meant
 - using strCat() when strAppend() was meant
 - using the remove_if() algorithm without passing the result to a
   follow-up erase() call
In the openMSX code I did not see any new warnings after adding the
annotations. But it's always good to know that this class of bugs is not
present in our code, and will not be accidentally added in the future. (Or just
catch the bug already during compile-time rather than during testing).

How/when to add annotations:
Short answer: I don't know. This is a new c++ feature and I don't have much
experience in using it. The only way to gain that experience is to actually
start using it. Hence this experiment.

In principle [[nodiscard]] makes sense on the majority of non-void functions.
(So you can reasonably argue that it would have been better to annotate
functions where it's OK to ignore the result. Unfortunately to keep the c++
language backwards compatibility that was not possible.) Though I'm not sure
yet it's a good idea to 'clutter' most code with [[nodiscard]]. Instead I'll
start by annotating functions that are:
- often called (e.g. utility functions or methods of important base classes)
- ignoring the result is always wrong (e.g. malloc()-like functions, or
  important error return codes)
This patch only changes the util/ subdirectory yet.