AviWriter: replace macros with lambda functions
Because lambdas are more type-safe than macros.

Found while looking for "invalid identifiers" in our code. These are
identifiers which either:
 1) contain a double underscore (__) somewhere
 2) start with an underscore followed by a capital letter
 3) start with an underscore and are defined at the global level
All such identifiers are reserved for the compiler or for the implementation of
standard libraries. Using them results in undefined behavior (although in
practice it often happens to work fine).

Before this patch the AviWriter code used the identifier "_S_" in some macros.
This matches the above criteria number 2. After this patch our code is clean(*)
wrt this stuff.

(*) The code still does contain many identifier of categories 1-3, but none
of these are defined by us. Some examples:
 1) __GNUC__, __m128, __restrict, __APPLE__
 2) _WIN32, _BitScanReverse64, _MM_SHUFFLE
 3) _wmkdir, _mm_load_si128