[5b/7] split MSXMemoryMapper class into 2
Before we had this class hierarchy:
  class MSXMemoryMapper
  class MusicalMemoryMapper : public MSXMemoryMapper
  class PanasonicRam : public MSXMemoryMapper

This patch changes that to:
  class MSXMemoryMapperBase
  class MSXMemoryMapper : public MSXMemoryMapperBase
  class MusicalMemoryMapper : public MSXMemoryMapperBase
  class PanasonicRam : public MSXMemoryMapperBase

Thus the old MSXMemoryMapper is split into an abstract base class and a
concrete leaf class. In general it's a good idea to avoid inheriting from
concrete (=non-abstract) classes.

This allows to change (parts of) the behavior of MSXMemoryMapper without also
changing the behavior of MusicalMemoryMapper and PanasonicRam. And that's
exactly what we need in the next patch in this series.

This patch already makes some other preparations: the method
MSXMemoryMapperBase::writeIO() is left pure virtual (=abstract). The old
implementation of this method is moved to a new (protected) method called
writeIOImpl(). All subclasses must implement writeIO() themselves now,
typically those implementations will call writeIOImpl() from the base class.

One additional change is that the cache line invalidation logic is now moved
from MSXMapperIO to all MSXMemoryMapperBase subclasses. For now all those
implementations are still the same (so it could still be implemented in the base
class), but later patches will refine this per device.