[5c/7] Don't invalidate on MSXMemoryMapper segment switch
Invalidating a cache line means setting it's status to 'unknown'. The internal
representation for 'unknown' is a nullptr. Later, in a lazy way, that cache
line will be filled-in again with a backdoor-pointer.

Though in case of a memory mapper segment, instead of invalidating the cache
line (= setting it to nullptr), we can immediately set it to the correct value.
This is the same effort as invalidating, but it saves work later on. So that's
what this patch does.

One complication is the interaction with the UMR-checker. When this checker is
in use we cannot allow reads/writes from/to (parts of) the memory mapper to be
bypassed. So we need some extra logic to check if pre-filling the cache lines
is allowed. But usually the UMR checker is disabled, and then this check only
has a very small impact on performance.

Here are the updated profile counters (compare to the results from patch [2/7]
in this series):

Zexall
  NonCachedRead: 691296
  NonCachedWrite: 154074
  GetReadCacheLine: 168
  GetWriteCacheLine: 27
  SlowRead: 691154
  SlowWrite: 154067
  DisallowCacheRead: 52698
  DisallowCacheWrite: 104479
  InvalidateAllSlots: 10
  InvalidateReadWrite: 28
  InvalidateRead: 0
  InvalidateWrite: 0
  FillReadWrite: 4
  FillRead: 0
  FillWrite: 0
 Only small improvements (a few 100 NonCachedRW less), but that's expected:
 once the cache lines for the memory mapper slot are filled in, they remain the
 same (the 4 FillReadWrite events are likely the 4 OUT instructions to ports
 0xfc-0xff done by the BIOS to initialize the mapper).

pa3
  NonCachedRead: 1965146
  NonCachedWrite: 178809
  GetReadCacheLine: 169
  GetWriteCacheLine: 34
  SlowRead: 1965009
  SlowWrite: 178801
  DisallowCacheRead: 61353
  DisallowCacheWrite: 120724
  InvalidateAllSlots: 20
  InvalidateReadWrite: 30
  InvalidateRead: 0
  InvalidateWrite: 0
  FillReadWrite: 44740
  FillRead: 0
  FillWrite: 0
 This program does more memory mapper segment switches (see the FillReadWrite
 events). This results in 1 or 2 orders of magnitude reduction in
 GetRWCacheLine events and in about 10000 less NonCachedRW events.

mgear2
  NonCachedRead: 2172417
  NonCachedWrite: 2482145
  GetReadCacheLine: 2150888
  GetWriteCacheLine: 159345
  SlowRead: 21551
  SlowWrite: 2482130
  DisallowCacheRead: 2814
  DisallowCacheWrite: 4924
  InvalidateAllSlots: 10
  InvalidateReadWrite: 665497
  InvalidateRead: 1684599
  InvalidateWrite: 0
  FillReadWrite: 4
  FillRead: 0
  FillWrite: 0
 Same as 'zexall': this test doesn't use the memory mapper (much), so only
 small improvements.