[4d/7] where possibly only invalidate 'read' or only 'write'
Instead of always invalidating both read and write cache lines.

For example after switching banks in a mega-rom, future reads will return a
different result. Thus read-lines (of a specific region) must be invalidated.
But writing to that region behaves the same as before: writes either continue
to be ignored (we call this an 'unmappedWrite' region) or continue to trigger
bank switches (emulated via a non-cacheable region).

It's sometimes tricky to see when exactly it's allowed to only invalidate read
or only write. The bad thing is that when you get this wrong, the code may
still work correctly in many cases (thus testing will not always reveal these
mistakes). Therefor this patch is relatively conservative in converting
read+write invalidation into read- or write-only. Later, maybe based on profile
results, we can improve this further.