[5d/7] CheckedRam bugfix UMR-detection
Discovered while working on the previous patch. This bug is also present on
master (I mean not introduced in this patch series), so I could have split-off
this patch from the series, but whatever.

This fix makes the UMR-checker a bit slower. But this is anyway only a debug
feature so it doesn't really matter.