[5e/7] Similar to patch '5c' but now for PanasonicRam
PanasonicRam implements the turboR memory mapper. This mapper differs from a
regular memory mapper because of the turboR DRAM mode. In DRAM mode part of the
memory mapper becomes read-only. This also affects the cache-line stuff.

In openMSX we emulate other MSX devices that implement memory mapper behavior.
In this patch series I've only optimized MSXMemoryMapper and PanasonicRam,
because these are the most frequently used. The others continue to be just as
fast as before the start of this patch series, but if we ever see them in a
performance profile, we have the option to also optimize those.