[6/7] Fill RomBlocks Rcache
Similar to memory mappers in the previous few patches: instead of simply
invalidating cache lines on a Mega-ROM bank-switch we can immediately fill in
the correct pointer value. Difference with a memory mapper: for ROM we only
need to fill in the read part (and leave the write part unchanged).

This patch only makes the change in the (templatized) RomBlocks base class.
Thus effectively for the Rom4kBBlocks, Rom8kBBlocks and Rom16kBBlocks base
classes. This covers most but not all of the ROM mappers that we emulate in
openMSX. So there is room for improvement later.

Here are the updated profile counters (compare to the results from patch [5c/7]
in this series):

Zexall and pa3: (as expected) unchanged
mgear2:
  NonCachedRead: 21683
  NonCachedWrite: 2482145
  GetReadCacheLine: 154
  GetWriteCacheLine: 159345
  SlowRead: 21551
  SlowWrite: 2482130
  DisallowCacheRead: 2814
  DisallowCacheWrite: 4924
  InvalidateAllSlots: 10
  InvalidateReadWrite: 665497
  InvalidateRead: 0
  InvalidateWrite: 0
  FillReadWrite: 4
  FillRead: 1684599
  FillWrite: 0
 There's roughly a factor 100x less NonCachedRead and 1000x less
 GetReadCacheLine events. The numer of corresponding write events remains the
 same. There are quite some SlowWrite events, most likely those are mega-ROM
 bank switches, those now lead to a large number of FillRead events. (Good:
 that shows this patch is working). There are still a large number of
 InvalidateReadWrite events. Those are triggered by enabling/disabling of the
 SCC registers, for that see the next patch in this series.