avoid UMR on savestate
Reported by 'p_gimeno' on IRC, thanks!

When building with clang this not only resulted in random data in the savestate
file, but even in a crash! This was because clang assumed booleans can only
have the value '0' or '1', it did some calculations on that value to determine
the length paramater for a memcpy(), but when the original assumption was wrong
(because of UMR) this resulted in a buffer overrun.