Fix bug#1201, segfault on unaligned memory access
Thanks to Emily82 for reporting and to mthuurne for analyzing the problem. I
merely fixed the problem, which was easy after the analysis.

See the bug report for a lot more details:
  https://github.com/openMSX/openMSX/issues/1201

I fully removed the OPENMSX_UNALIGNED_MEMORY_ACCESS constant (and the
infrastructure to calculate this). Using this is too dangerous (it might work
for years, but breaks with a newer smarter compiler). Also usually it isn't
needed anymore to get good performance (with modern compilers).