Implement structured bindings support gl::vec
Inspired by this "C++ meeting 2019" lightningtalk:
  https://www.youtube.com/watch?v=YC_TMAbHyQU&t=219s

The main purpose of the 'gl::vec' class is to group a bunch of 'scaler'
variables together into 1 'vector' variable. For example 3 floats x, y, and z
together form a 3D-position. So we can replace code like:
  float x1 = ..., x2 = ...;
  float y1 = ..., y2 = ...;
  float z1 = ..., z2 = ...;
  x1 += x2;
  y1 += y2;
  z1 += z2;
with:
  vec3 pos1 = ..., pos2 = ...;
  pos1 += pos2;

Access to the individual scalar elements in the vector is possible via
operator[]. But when possible prefer to avoid this.

However in some cases it is unavoidable. For example when calling a C-function
that has a scalar interface:
  vec2 pos = getPosition();
  vec2 size = getSize();
  glViewport(pos[0], pos[1], size[0], size[1]);

With this patch that code can be simplified to:
  auto [x, y] = getPosition();
  auto [w, h] = getSize();
  glViewport(x, y, w, h);