Fix possible crash on invalid command line
Reported by p_gimeno on IRC. To reproduce execute e.g.:
   openmsx -model msx1

Because '-model is not recognized as a valid command line option, openMSX will
try to interpret it as a filename. As part of that it will check the filename
extension.

To get the filename extension we had code like:
  extension = FileOperations::getExtension(arg).substr(1);

The return value of getExtension() is specified to include the leading '.'
character. This is similar to the standard c++17 function
std::filesystem::path::extension() (but we can only start using that function
once we require gcc-8). The call to string_view::substr() will then remove that
'.' character.

When that code was originally written we had our own implementation of
string_view, because at the time we couldn't use the standard c++17
std::string_view class yet. Later we did switch to std::string_view.
Unfortunately our version differed slightly from the standard behavior when the
value of the 'pos' parameter (1st parameter) of subst() was larger than the
size of the string. In our old version this resulted in an empty string, but
with std::string this will throw an out_of_range exception.

Easily fixed once the problem was identified.

I will now audit all other uses of std::string_view::substr(), maybe there are
other places where we rely on this non-standard behavior.