Change Coleco memory initialization to fix The Heist
We set it to all zeros, which seems to work for ColEm.

EDIT by Manuel: we don't know for sure whether this is the correct pattern or
whether it is always this pattern. But it surely fixes The Heist and we haven't
heard about The Heist not working on real hardware, nor has Pedro found ROMs
that stopped working because of this patch.