Rename member 'name' of Completer class.
This internal (private) member causes a shadow warning if used as parameter
name ("name") in members of derived classes. So, I'm sacrificing the beauty of
the name of this private internal member to make the names of parameters in
derived classes easier to read and understand.

Fixes a warning introduced by my previous changes.