small cleanup in SDLVisibleSurface/SDLOutputSurface
In SDLVisibleSurface:
Directly use the SDLVisibleSurface::surface member variable, instead of
querying (the same value) from the base class via
SDLOutputSurface::getSDLSurface(). As preparation to (possibly in the future)
remove this from the base class.

In SDLOutputSurface:
Use the getter method getSDLSurface() instead of direct access to the surface
member variable. To make it more clear that, for now, calls to setSDLSurface()
(by some subclasses) are still required. (I hope to remove those later).