Disable V99x8 specific debug writes on TMS99xx
Internally TMS99xx (MSX1) VDP emulation is using the V99x8 VDP emulation code
but with some features disabled. Most notably writes to register 8 and above.
(Ignoring various details).

Though debug writes to those register were NOT ignored, but because of
implementation details (masking), bits in those registers could only be reset
(not set). Fixed by really ignoring debug writes to those registers. An
alternative fix could be to only expose 8 (instead of 64) registers in the
debug interface. But unfortunately that will break some existing Tcl scripts
(e.g. a script that checks R#9 bit 1 to query PAL or NTSC).

Similarly on TMS99xx the palette was exposed via the debug interface. (But the
code was not prepared to handle a changing TMS palette). Fixed by ignoring
debug writes on TMS99xx. Again an alternative fix is to not expose the palette
debuggable, but having a read-only palette could be useful anyway.