Update Android ARM compiler name
This matches the name in NDK 20.

It also ups the API level from 16 to 18 (Android 4.3), to work around
an issue in SDL2 where <cpu-features.h> is not found, since that is
not a system header but a header of a library that's included in the
NDK but not in the default search path. On API level 18, autoconf
will define HAVE_GETAUXVAL and the header won't be included by SDL2,
since it will get the info using the getauxval() call instead.