Fix issue#1217: Assertion failed (frame <= time) in VDPAccessSlots.hh
TBH I have no idea why this issue didn't trigger in a lot more test cases. I
don't think there's anything special in #1217 that causes the assert to
trigger. My best guess is (un)lucky timing??

The problem was that access-slot calculations need some reference time point.
We use the start of the frame for that. But the calculations don't expect that
point to be in the future. And therefor we need to synchronize with the
VDPCmdEngine at the end of the frame. Before this patch that sync was missing.