Add another workaround for Windows 10 symlinks
In d690b5a0 (Handle symlink creation failing on Windows 10) I added
a workaround that print a warning if a symlink couldn't be created
when unpacking a tarball. But we also create symlinks during
installation, if the os.symlink() function exists.

Instead of a warning, I opted here to not attempt to create symlinks
on any OS from the Windows NT family. That avoids noise in the logs.