Fix regression introduced in f5be4ee946f2d2
The built-in software on a turbor-ST no longer worked.
Thanks to p_gimeno for reporting and even bisecting the problem!

Commit f5be4ee946f2d2 enabled R-cache-line filling on bank-switch in the
'RomBlocks' base class. That turned out to be overly ambitious because for
(some of the) subclasses that override the getReadCacheLine() method, this
auto-fill behavior was not (always) correct. A more prudent approach would have
been to only enable it for specific (audited) subclasses.

Anyway, damage is done. This change does bring some nice performance
improvements, so I'd prefer to keep that. Therefor, instead of (partly)
reverting f5be4ee946f2d2, I audited all RomBlocks subclasses (that override
getReadCacheLine()) and fixed the incompatibilities.