Change YM2413Core (virtual) interface to prepare for NukeYKT core
Rename writeReg() to pokeReg(), this should now only be used by the debugger.

To actually write registers, use the new method writePort(). Like on a real
YM2413 you must first write the register number to the register port, followed
by writing the register value to the data port.

Before this process was abstracted away (because the existing YM2413 cores
didn't care about this, and then also this process could be shared among the
different cores). Though the more accurate NukeYKT YM2413 core does take the
timing of write to the register/data port into account.

TODO this patch extends the interface but does not yet provide accurate
'sub-sample' timing information. That's for the next patch.