Expose 'EmuClock' in the 'ResampleSoundDevice' class
Each 'ResampleSoundDevice' object already contained 'ResampleAlgo' object. And
most of those already kept track of the emulated time (only exception was
'ResampleTrivial') via a 'DynamicClock' member variable. This patch moves that
clock from the 'ResampleAlgo'-subclasses to the 'ResampleSoundDevice' class.

In a follow-up patch we'll use that clock to pass sub-sample timing information
to the YM2413Core class.