Added NukeYKT YM2413 emulation core
To allow easier experimentation I've also made this new core the default. I'm
not sure it should also be the default in the next release. Probably it should,
and eventually we may even drop the old YM2413 cores. But it definitely needs
more testing.

To choose between the different YM2413 cores you must edit the XML config file
(machine or extension). Inside the <FMPAC> or <MSX-MUSIC> tag add a new tag
    <alternative>...</alternative>
The possible values for '...' are:
- 'NukeYKT'
   This is the current default.
- 'Original-NukeYKT'
   This is the original (unmodified) NukeYKT code, more on this below.
- 'Okazaki'
   This was the default before this patch.
- 'Burczynski'
   Even before this patch we had an alternative YM2413 implementation.

I've added both the original unmodified NukeYKT code and a more optimized
version. Both generate identical sound, but the optimized version runs more
than 3x faster (on average). I've tested/verified this on all the YM2413 vgm
files I could find (378 files).

I'm not sure I should have added the 'Original-NukeYKT' core. For sure it's a
debug only feature. And we should remove or at least disable it in a release.
Maybe a better debug feature is to run both versions simultaneously and assert
that they indeed generate identical output.