fix bw-compat with old savestates after 4ea01c915
When I added the YM2413 NukeYKT cores, I changed the default YM2413 core (when
nothing is specified in the hardware config file) from Okazaki to NukeYKT. This
broke backwards compatibility with old savestates. (Because when the config in
that savestate did not explicitly select a core, it would instantiate a NukeYKT
core, but the serialized data in the savestate was for a Okazaki core).

To fix this I reverted the default to Okazaki and instead explicitly select the
NukeYKT core in all our machine/extension configs.

When in the future we drop support for the Okazaki/Burczynski cores, we can then
add transition code for old savestates. (A perfect transition will be
impossible because the internals of those cores differ too much, but maybe only
restoring the register values is already acceptable).