2nd attempt to fix debian bug 955551
This is a follow-up on 0c259566e5f which tried to fix
    https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=955551

Now I actually tried to compile on s390x myself. By using clang with
option '--target=s390x-linux-gnu' on godbolt:
    https://godbolt.org/z/y8JwZD

Initially I got a very similar error as in the debian bug report. After a
(small) fix it compiled fine.