Fix YM2413NukeYKT savestate
The root cause was an uninitialized 'uint8_t regs[64]' array. This is not part
of the original NukeYKT core, but we added it to be able to debug-read the
registers. So for normal sound generation there was no problem.

However, to reduce redundancy in savestates, we try to reduce internal state
from register values. And then it does become a problem when those register
value are wrong (uninitialized). (Savestate is also not part of the original
NukeYKT core).