Fix RUN command being typed while in menu in some models
Some models use the H_READ hook for displaying the menu, and since the old code used this hook to type, the text was incorrectly typed while in the menu.

The H_MAIN hook is executed in many occasions, the first one being right after H_READ, and indicates that the interpreter is ready. This doesn't happen during the menus, therefore a viable strategy is to type the text on the first call to H_MAIN after an H_READ.